pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
description = "Multilingual voice calling agent using LiveKit, Groq, and Piper TTS"
readme = "README.md"
authors = [
    { name = "AimanSiddiqui99", email = "aiman.siddiqui@whiteboxtech.net" }
]
requires-python = ">=3.12"
dependencies = [
    "livekit-agents~=1.3",
    "livekit-plugins-groq~=1.2",
    "livekit-plugins-silero~=1.3",
    # "livekit-plugins-turn-detector~=1.3",  # disabled: OOM kill on Railway free tier
    "piper-tts>=1.2",
    "onnxruntime>=1.19,<1.24",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "faster-whisper>=1.0",
    "numpy>=1.24",
    "qdrant-client>=1.11",
    "fastembed>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[project.scripts]
livekit-voice-agent = "livekit_voice_agent:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
